% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{block_chart_type_css}
\alias{block_chart_type_css}
\title{Control chart type buttons via CSS}
\usage{
block_chart_type_css()
}
\value{
HTML style tag with CSS to control chart type buttons
}
\description{
Control chart type buttons via CSS
}
