% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-block-utils.R
\name{block_container_script}
\alias{block_container_script}
\title{Generate container query script for responsive blocks}
\usage{
block_container_script()
}
\value{
HTML script tag
}
\description{
Sets up container queries if supported by the browser.
}
