% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.ggplot-package}
\alias{blockr.ggplot}
\alias{blockr.ggplot-package}
\title{blockr.ggplot: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Nicolas Bennett \email{nicolas@cynkra.com}

}
\keyword{internal}
