% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-block.R
\name{build_theme_choices}
\alias{build_theme_choices}
\title{Build theme choices list based on available packages}
\usage{
build_theme_choices()
}
\value{
A named list of theme choices for selectInput
}
\description{
Build theme choices list based on available packages
}
\keyword{internal}
