% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-blockr.R
\name{theme_blockr}
\alias{theme_blockr}
\title{A custom blockr ggplot2 theme}
\usage{
theme_blockr()
}
\description{
This theme is based on \code{hrbrthemes::theme_ipsum()} which is issued under an
MIT license.
}
