% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.ggplot-package}
\alias{blockr.ggplot}
\alias{blockr.ggplot-package}
\title{blockr.ggplot: Interactive 'ggplot2' Visualization Blocks}
\description{
Extends 'blockr.core' with interactive blocks for data visualization using 'ggplot2'. Users can build charts through a graphical interface without writing code directly. Includes PowerBI-style chart types (bar charts, line charts, pie charts, scatter plots) as well as statistical plots (boxplots, histograms, density plots, violin plots, heatmaps) with rich customization options and intuitive user interfaces.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bristolmyerssquibb.github.io/blockr.ggplot/}
  \item Report bugs at \url{https://github.com/BristolMyersSquibb/blockr.ggplot/issues}
}

}
\author{
\strong{Maintainer}: Christoph Sax \email{christoph@cynkra.com} (\href{https://orcid.org/0000-0002-7192-7044}{ORCID})

Authors:
\itemize{
  \item Nicolas Bennett \email{nicolas@cynkra.com}
  \item David Granjon \email{david@cynkra.com}
  \item Mike Page \email{mike@cynkra.com}
}

Other contributors:
\itemize{
  \item  Bristol Myers Squibb [funder]
}

}
\keyword{internal}
