% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.ggplot-package}
\alias{blockr.ggplot}
\alias{blockr.ggplot-package}
\title{blockr.ggplot: Interactive ggplot2 Charts for Blockr}
\description{
Provides comprehensive blocks for plotting with blockr using 'ggplot2'. Includes PowerBI-style chart types (bar charts, line charts, pie charts, scatter plots) as well as statistical plots (boxplots, histograms, density plots, violin plots, heatmaps) with rich customization options and intuitive user interfaces.
}
\author{
\strong{Maintainer}: Nicolas Bennett \email{nicolas@cynkra.com}

Authors:
\itemize{
  \item Christoph Sax \email{christoph@cynkra.com}
}

}
\keyword{internal}
