% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_block.R
\name{new_boxplot_block}
\alias{new_boxplot_block}
\title{Boxplot block constructor}
\usage{
new_boxplot_block(x = character(), color = character(), ...)
}
\arguments{
\item{x, color}{Columns to use for \code{\link[ggplot2:aes]{ggplot2::aes()}}}

\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}
}
\description{
This block draws a boxplot using \code{\link[ggplot2:geom_boxplot]{ggplot2::geom_boxplot()}}.
}
